requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.28.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.6
defusedxml==0.7.1
distro==1.9.0
Django==5.0.6
django-allauth==0.63.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.1
filelock==3.15.1
frozenlist==1.4.1
fsspec==2024.6.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.7
inflection==0.5.1
jiter==0.4.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-anthropic==0.1.15
langchain-community==0.2.1
langchain-core==0.2.5
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langgraph==0.0.66
langsmith==0.1.77
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==1.26.4
openai==1.30.2
orjson==3.10.3
packaging==23.2
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
PyMySQL==1.1.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
tavily-python==0.3.3
tenacity==8.3.0
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
